Extract App component
Extract App component.
